Add a branch with some code to compute the closest facet in a triangular mesh to a given point in 3D.
The code builds and works, but even though it finds the closest facet, I do
not know how to get an index to it. That is, the code finds the closest facet
as some kind of type Primitive, and I don't know how to tell that this is e.g.
facet 8 in the "tri" mesh description.

The documentation in CGAL is not very good, and I had to give up. I have coded a
not very efficient alternative in Matlab, closest_trifacet() in the ManifoldToolbox
and will use that for the moment.

I also emailed the cgal-discuss mailing list, but got no reply so far

http://cgal-discuss.949826.n4.nabble.com/Getting-index-of-facet-closest-to-a-point-td4657635.html